Split omemo/utils.js into headless and UI versions
This replaces 'plugins/omemo' with 'plugins/omemo-views' and
'headless/plugins/omemo'.
