+++ src/headless/plugins/omemo/consts.js
