+++ src/headless/plugins/omemo/device.js
