+++ src/headless/plugins/omemo/devicelist.js
