+++ src/headless/plugins/omemo/utils.js
@@ -2,19 +2,12 @@
 import concat from 'lodash-es/concat';
 import difference from 'lodash-es/difference';
 import log from '@converse/headless/log';
-import tpl_audio from 'templates/audio.js';
-import tpl_file from 'templates/file.js';
-import tpl_image from 'templates/image.js';
-import tpl_video from 'templates/video.js';
 import { KEY_ALGO, UNTRUSTED, TAG_LENGTH } from './consts.js';
 import { MIMETYPES_MAP } from 'utils/file.js';
 import { __ } from 'i18n';
 import { _converse, converse, api } from '@converse/headless/core';
-import { html } from 'lit';
+import { getURI } from '@converse/headless/utils/url.js';
 import { initStorage } from '@converse/headless/utils/storage.js';
-import { isError } from '@converse/headless/utils/core.js';
-import { isAudioURL, isImageURL, isVideoURL, getURI } from '@converse/headless/utils/url.js';
-import { until } from 'lit/directives/until.js';
 import {
     appendArrayBuffer,
     arrayBufferToBase64,
@@ -25,7 +18,7 @@ import {
     stringToArrayBuffer
 } from '@converse/headless/utils/arraybuffer.js';
 
-const { Strophe, URI, sizzle, u } = converse.env;
+const { Strophe, sizzle, u } = converse.env;
 
 export function formatFingerprint (fp) {
     fp = fp.replace(/^05/, '');
@@ -197,36 +190,26 @@ async function getAndDecryptFile (uri) {
     }
 }
 
-function getTemplateForObjectURL (uri, obj_url, richtext) {
-    if (isError(obj_url)) {
-        return html`<p class="error">${obj_url.message}</p>`;
-    }
-
-    const file_url = uri.toString();
-    if (isImageURL(file_url)) {
-        return tpl_image({
-            'src': obj_url,
-            'onClick': richtext.onImgClick,
-            'onLoad': richtext.onImgLoad
-        });
-    } else if (isAudioURL(file_url)) {
-        return tpl_audio(obj_url);
-    } else if (isVideoURL(file_url)) {
-        return tpl_video(obj_url);
-    } else {
-        return tpl_file(obj_url, uri.filename());
-    }
-
+export const omemo = {
+    decryptMessage,
+    encryptMessage,
+    formatFingerprint
 }
 
-function addEncryptedFiles(text, offset, richtext) {
+export function processEncryptedFiles (text) {
     const objs = [];
     try {
         const parse_options = { 'start': /\b(aesgcm:\/\/)/gi };
         URI.withinString(
             text,
             (url, start, end) => {
-                objs.push({ url, start, end });
+                const uri = getURI(text.slice(o.start, o.end));
+                objs.push({
+                    uri,
+                    start,
+                    end,
+                    obj_url: getAndDecryptFile(uri); // this is a promise
+                });
                 return url;
             },
             parse_options
@@ -235,21 +218,8 @@ function addEncryptedFiles(text, offset, richtext) {
         log.debug(error);
         return;
     }
-    objs.forEach(o => {
-        const uri = getURI(text.slice(o.start, o.end));
-        const promise = getAndDecryptFile(uri)
-            .then(obj_url => getTemplateForObjectURL(uri, obj_url, richtext));
 
-        const template = html`${until(promise, '')}`;
-        richtext.addTemplateResult(o.start + offset, o.end + offset, template);
-    });
-}
-
-export function handleEncryptedFiles (richtext) {
-    if (!_converse.config.get('trusted')) {
-        return;
-    }
-    richtext.addAnnotations((text, offset) => addEncryptedFiles(text, offset, richtext));
+    return objs;
 }
 
 /**
@@ -306,23 +276,16 @@ export function onChatBoxesInitialized () {
     });
 }
 
-export function onChatInitialized (el) {
-    el.listenTo(el.model.messages, 'add', message => {
+export function onChatBoxInitialized(model) {
+    model.listenTo(model.messages, 'add', message => {
         if (message.get('is_encrypted') && !message.get('is_error')) {
-            el.model.save('omemo_supported', true);
+            model.save('omemo_supported', true);
         }
     });
-    el.listenTo(el.model, 'change:omemo_supported', () => {
-        if (!el.model.get('omemo_supported') && el.model.get('omemo_active')) {
-            el.model.set('omemo_active', false);
-        } else {
-            // Manually trigger an update, setting omemo_active to
-            // false above will automatically trigger one.
-            el.querySelector('converse-chat-toolbar')?.requestUpdate();
+    model.listenTo(model, 'change:omemo_supported', () => {
+        if (!model.get('omemo_supported') && model.get('omemo_active')) {
+            model.set('omemo_active', false);
         }
-    });
-    el.listenTo(el.model, 'change:omemo_active', () => {
-        el.querySelector('converse-chat-toolbar').requestUpdate();
     });
 }
 
@@ -330,6 +293,7 @@ export function getSessionCipher (jid, id) {
     const address = new libsignal.SignalProtocolAddress(jid, id);
     return new window.libsignal.SessionCipher(_converse.omemo_store, address);
 }
+
 
 function getJIDForDecryption (attrs) {
     const from_jid = attrs.from_muc ? attrs.from_real_jid : attrs.from;
@@ -458,6 +422,7 @@ export function addKeysToMessageStanza (stanza, dicts, iv) {
     }
     return Promise.resolve(stanza);
 }
+
 
 /**
  * Given an XML element representing a user's OMEMO bundle, parse it
@@ -676,6 +641,7 @@ export async function initOMEMO (reconnecting) {
     api.trigger('OMEMOInitialized');
 }
 
+
 async function onOccupantAdded (chatroom, occupant) {
     if (occupant.isSelf() || !chatroom.features.get('nonanonymous') || !chatroom.features.get('membersonly')) {
         return;
@@ -709,73 +675,6 @@ async function checkOMEMOSupported (chatbox) {
         chatbox.set('omemo_active', true);
     }
 }
-
-function toggleOMEMO (ev) {
-    ev.stopPropagation();
-    ev.preventDefault();
-    const toolbar_el = u.ancestor(ev.target, 'converse-chat-toolbar');
-    if (!toolbar_el.model.get('omemo_supported')) {
-        let messages;
-        if (toolbar_el.model.get('type') === _converse.CHATROOMS_TYPE) {
-            messages = [
-                __(
-                    'Cannot use end-to-end encryption in this groupchat, ' +
-                        'either the groupchat has some anonymity or not all participants support OMEMO.'
-                )
-            ];
-        } else {
-            messages = [
-                __(
-                    "Cannot use end-to-end encryption because %1$s uses a client that doesn't support OMEMO.",
-                    toolbar_el.model.contact.getDisplayName()
-                )
-            ];
-        }
-        return api.alert('error', __('Error'), messages);
-    }
-    toolbar_el.model.save({ 'omemo_active': !toolbar_el.model.get('omemo_active') });
-}
-
-export function getOMEMOToolbarButton (toolbar_el, buttons) {
-    const model = toolbar_el.model;
-    const is_muc = model.get('type') === _converse.CHATROOMS_TYPE;
-    let title;
-    if (model.get('omemo_supported')) {
-        const i18n_plaintext = __('Messages are being sent in plaintext');
-        const i18n_encrypted = __('Messages are sent encrypted');
-        title = model.get('omemo_active') ? i18n_encrypted : i18n_plaintext;
-    } else if (is_muc) {
-        title = __(
-            'This groupchat needs to be members-only and non-anonymous in ' +
-                'order to support OMEMO encrypted messages'
-        );
-    } else {
-        title = __('OMEMO encryption is not supported');
-    }
-
-    let color;
-    if (model.get('omemo_supported')) {
-        if (model.get('omemo_active')) {
-            color = is_muc ? `var(--muc-color)` : `var(--chat-toolbar-btn-color)`;
-        } else {
-            color = `var(--error-color)`;
-        }
-    } else {
-        color = `var(--muc-toolbar-btn-disabled-color)`;
-    }
-    buttons.push(html`
-        <button class="toggle-omemo" title="${title}" data-disabled=${!model.get('omemo_supported')} @click=${toggleOMEMO}>
-            <converse-icon
-                class="fa ${model.get('omemo_active') ? `fa-lock` : `fa-unlock`}"
-                path-prefix="${api.settings.get('assets_path')}"
-                size="1em"
-                color="${color}"
-            ></converse-icon>
-        </button>
-    `);
-    return buttons;
-}
-
 
 async function getBundlesAndBuildSessions (chatbox) {
     const no_devices_err = __('Sorry, no devices found to which we can send an OMEMO encrypted message.');
@@ -857,10 +756,4 @@ export async function createOMEMOMessageStanza (chat, data) {
     stanza.c('store', { 'xmlns': Strophe.NS.HINTS }).up();
     stanza.c('encryption', { 'xmlns': Strophe.NS.EME,  namespace: Strophe.NS.OMEMO });
     return { message, stanza };
-}
-
-export const omemo = {
-    decryptMessage,
-    encryptMessage,
-    formatFingerprint
 }
