+++ src/plugins/omemo-views/fingerprints.js
